Changed No Storage/No Materials lists to not clear when map is updated.
There was an issue where tasks that didn't have a path to a storage or
materials would end up on the no storage/no materials lists, even if
these things did exist and there was just no path to them. The fix for
this was to clear the no storage/no material lists whenever a map update
happened. This was inefficient, and has been changed. Now the behavior
calculator will check to see if the storage or item exists on the map,
and if it does, will report the task as No Path Found.