(wip) Added new texture states to detect if adjacent tiles are a specific state.
This is meant to handle jagged corners between packed dirt blocks. It's
a bit more complicated than first believed, because if the texture it
relies on isn't calculated first, it will return the wrong texture.
Textures that depend on other textures should probably have a separate
state list, to avoid this problem.