Fixed bug where golems would finish task even if a building was deleted.
For some tasks that didn't register as "active behaviors" for a
building, deleting the building mid-task would not cause the golem to
fail the behavior (ie, if a golem went into a house to repair itself,
and the house was deleted, the golem would finish the task). This
behavior ends correctly now, although golems do have a short pause after
this happens that needs to be investigated.