WIP (Issue #32)  GUI: Storage Menu
Pressing space will now switch to another menu that tells you how much
of each resource you have. The menu is not done, but it does give the
necessary information. Otherwise, it seems that update might run if the
menu isn't selected. This needs to be investigated, and if true, needs
to be fixed.