(Issue #40) Ramps that transform into Clay Blocks shouldn't use the 'unbuilt' texture.
Transform now will call constructionComplete() if the new building it
transforms into has a build time of zero. This also corrected a bug that
would replace the generic building with a copy of itself regardless of
whether or not a transformation happened. It is unknown if this is
related to the bug ramps have right now (since ramps are the only tile
that transform, at the moment).