Fixed bug not adding extra weight; not properly calculating max golem population.
There was a bug returning all added weight because of priority (ie,
building an Obelisk if low on mana and no Obelisks) as 0, and a bug
calculating the maximum CLAY golem population as the city's entire golem
population compared to the number of houses, instead of CLAY golem
population compared to houses. Both of these bugs have been fixed.