(Issue #90) Skyways, Pressers and Spinners (Mana/Architecture GOAL #4)
Skyways, Pressers, and Spinners all added. Skyways are functional
bridges, that can be used to build over supports. Pressers produce
coolant. Spinners are currently useless. Note, deconstructing a skyway
leads to the unfortunate situation where a clay block can be placed over
a support. This needs to be resolved in (Issue #91).