Massive overhaul to GBP; everything now managed on one thread, while EventBus thread queues actions for the main thread.
In an effort to avoid thread-related bugs, the main thread and the
EventBus thread no longer modify the same lists. Instead, EventBus
queues actions for the GBP to handle, and GBP does all the
list-modifying.

This also marks the correct of a bug with conversions, where conversion
behaviors added in batch would all have the same 'conversion' building
to claim. This caused god knows what problems, since behaviors would try
to claim buildings unrelated to them. There's a good chance this is why
(Issue #97) was happening, so until that occurs again, that issue will
be closed.

This also resolves (Issue #88) and (Issue  #102).